fix: update demo apps for ADR-002/ADR-010 native Swift types and API changes
iOS:
- Update DecodedAudioInfo to use native Int types instead of Int32
- Remove unnecessary Float() casts in SingafterRepository
- Add max(0, ...) clamp for floating-point precision in MelodyEvalViewModel
- Update NoteEval, Effects, Singafter ViewModels for new API patterns

Android:
- Add SingalongViewModel and SingalongModels for MVVM architecture
- Add SingafterModels for cleaner data structures
- Update Singalong and Singafter views and viewmodels

+++ demo-apps/ios/VoxaTraceDemo/VoxaTraceDemo/Sections/Decoding/DecodingView.swift
@@ -173,7 +173,7 @@ struct DecodingView: View {
                 decodedInfo = DecodedAudioInfo(
                     sampleRate: result.sampleRate,
                     channels: result.numChannels,
-                    durationMs: Int64(result.durationMilliSecs),
+                    durationMs: result.durationMilliSecs,
                     dataSize: Int(result.audioData.size)
                 )
                 status = "Decoded successfully"
@@ -222,9 +222,9 @@ struct DecodingView: View {
 }
 
 struct DecodedAudioInfo {
-    let sampleRate: Int32
-    let channels: Int32
-    let durationMs: Int64
+    let sampleRate: Int
+    let channels: Int
+    let durationMs: Int
     let dataSize: Int
 }
